[Calendar] Fixed handling of startDate and endDate.
Issue: TNEF-4351

The platform expects that both start date and end date will be of the same type.
If only one of these values is set in JS, that value should be used as both start and end date, to ensure that they are properly filled for platform.
Note: this might be only related to isAllDay value set to true.

[Verification] TCT pass rate did not change.

Change-Id: I9c21e7d301be8df42471e060b1de5bdff180dc49
Signed-off-by: Pawel Andruszkiewicz <p.andruszkie@samsung.com>